+++ datafusion/physical-optimizer/tests/mod.rs
@@ -15,24 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-// Make cheap clones clear: https://github.com/apache/datafusion/issues/11143
-#![deny(clippy::clone_on_ref_ptr)]
-
-pub mod aggregate_statistics;
-pub mod coalesce_batches;
-pub mod combine_partial_final_agg;
-pub mod enforce_distribution;
-pub mod enforce_sorting;
-pub mod join_selection;
-pub mod limit_pushdown;
-pub mod limited_distinct_aggregation;
-mod optimizer;
-pub mod output_requirements;
-pub mod pruning;
-pub mod sanity_checker;
-pub mod topk_aggregation;
-pub mod update_aggr_exprs;
-pub mod utils;
-
-pub use optimizer::PhysicalOptimizerRule;
-pub mod utils;
+mod enforce_sorting;
+mod limited_distinct_aggregation;
+mod replace_with_order_preserving_variants;
+mod sanity_checker;
+mod test_utils;
